requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
pyodbc==5.0.1
selenium==4.15.2
python-dotenv==1.0.0
requests==2.31.0
webdriver-manager==4.0.1
openpyxl==3.1.2
